requirements.txt
azure-eventhub-checkpointstoreblob==1.1.1
azure-eventhub-checkpointstoreblob-aio==1.1.1
azure-eventhub==5.2.0
azure-servicebus==7.0.0
azure-storage-blob==12.6.0
azure-identity==1.5.0
azure-mgmt-eventhub==8.0.0
psutil==5.7.3
